Simplify struct symbol and correct value for offset field
All the character arrays are read only.  To make things more compact
and avoid hard-limits on the path or symbol name converted the the
fields in struct symbol.  It should have the added advatange that the
compiler can factor out all the identical path fields.

The offset field is an offset from the beginning of the file. The
"objdump -t" provides the vma address which is not the same value. The
vma address needs to be converted into a fileoffset.  The way this is
currently implemented will be wrong on some architectures and will
need to be fixed in a future commit.